fix #43, daemon child is a different binary
Fixed using start alias for the daemon and harcoding python as the
exec binary

BACKGROUND:
start-stop-daemon checks that the pid in the pid file matches the
binary stated in --exec (to prevetn pid reuse).

In our daemon, we don't fork and detach but start a different binary
as a detached child.

This confused start-stop-daemon, thinking the PID setted never
match the running program, ergo it's never running.

This was quite dangerous because starting multiple times lead to
overwrite the pidfile with the new child process which ended because
the port was in use. This leaves an invalid pid and makes the stop
function now unusable (pid not match runnig process)